Load Sentry browser tracing ASAP to link traces between frontend and backend
Otherwise @apollo/client was loaded beforehand and its instrumentation
didn't work
